适配 XBOX (#1280)
* 创建 xbox 的基本导航框架

* 简单处理排行榜页

* 使用新logo填补右上角空缺

* 添加xbox主视图内容

* 完成搜索页面

* 更新收藏页面

* 添加播放页

* 完成PGC和直播播放

* 添加简单的快捷操作

* 调整上下文菜单

* 更新应用介绍

+++ src/Models/Models.BiliBili/Live/LiveRoomCard.cs
@@ -15,7 +15,7 @@ namespace Bili.Models.BiliBili
         /// 直播间Id.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "id", Required = Required.Default)]
-        public int RoomId { get; set; }
+        public long RoomId { get; set; }
 
         /// <summary>
         /// 所属分区Id.
@@ -121,7 +121,7 @@ namespace Bili.Models.BiliBili
                 /// 标识符.
                 /// </summary>
                 [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "id", Required = Required.Default)]
-                public int Id { get; set; }
+                public long Id { get; set; }
 
                 /// <summary>
                 /// 名称.
